fix: seperate type declaration on color values
